Excel-Excel Version 0.0
Python Files ->

created excel_to_excel.py:

Initial version of code. The purpose of this file is to move information from excel files with consistent formatting to a destination excel file and organize it row by row. The code makes use of xlwings to read and write to excel documents.

In this version, a setup of variables. This is a temporary build and will later be replaced with a user friendly interface.

created building_blocks.py:

This file has important functions that are used by excel_to_excel.  Currently there are two functions - find_files and move_cell.

find_files searches through a directory and creates an array with the paths to each file with a key word.

move_cell moves a cell value to a destination cell

Excel Files ->

Destination.xlsx: The destination excel file

Test Case 1.xlsx A source excel file

Test Case 2.xlsx Another source excel file