Created new GUI
New gui.py was created. Four pages as of right now. The first page is a dashboard but only with a create button.

Page 2 is a browse file page. This is the page where you select the source directory. --Needs a way to remove files.

Page 3 is a page to select sheets which is dynamically added. The user will also select what to do with the sheet. --Needs updated ui

Page 4 will depend on the type selected by the user. Currently there is only a cell page with a drag and drop button.